+++ object_detectors/detectron2_lmi/utils/det_utils.py
@@ -1,13 +1,14 @@
 from detectron2 import model_zoo
 from detectron2.config import get_cfg
-from detectron2.data.datasets import register_coco_instances
 from detectron2.engine import DefaultTrainer
 import argparse
 import os
 from datetime import date
 import yaml
 from typing import Any, Callable, Dict, IO, List, Union
+from detectron2.utils.logger import setup_logger
 
+logger = setup_logger()
 
 def merge_a_into_b(a: Dict[str, Any], b: Dict[str, Any]) -> None:
     """based from fvcore/common/config.py"""
@@ -20,25 +21,6 @@ def merge_a_into_b(a: Dict[str, Any], b: Dict[str, Any]) -> None:
             merge_a_into_b(v, b[k])
         else:
             b[k] = v
-
-def register_datasets(dataset_dir: str, dataset_name: str):
-    """
-    Register the train and test datasets with Detectron2
-    """
-    if os.path.exists(dataset_dir):
-        annot_file = os.path.join(dataset_dir, "annotations.json")
-        images_path = os.path.join(dataset_dir, "images")
-        if os.path.isfile(annot_file) and os.path.isdir(images_path):
-            print(f"Registering dataset {dataset_name} from {dataset_dir}")
-            register_coco_instances(
-                dataset_name,
-                {},
-                annot_file,
-                images_path,
-            )
-        else:
-            raise ValueError(f"Invalid dataset directory {dataset_dir} for dataset {dataset_name}")
-    return None
 
 def create_config(cfg_file_path):
     # get the default config
@@ -71,13 +53,17 @@ def create_config(cfg_file_path):
     while os.path.exists(output_dir):
         version += 1
         output_dir = f"{date.today()}-v{version}"
+    
+    
+    cfg.OUTPUT_DIR = os.path.join(cfg.OUTPUT_DIR, output_dir)
 
-    cfg.OUTPUT_DIR = output_dir
+    logger.info(f"Output directory: {cfg.OUTPUT_DIR}")
+
 
     # create the output directory if it does not exist
 
-    if not os.path.exists(output_dir):
-        os.makedirs(output_dir)
+    if not os.path.exists(cfg.OUTPUT_DIR):
+        os.makedirs(cfg.OUTPUT_DIR)
 
     # save the config to the output directory
     cfg_file = os.path.join(output_dir, "config.yaml")
